YoloV5_Streamlit/requirements.txt
streamlit==1.10.0
torch==1.11.0
torchvision==0.12.0
numpy==1.21.5
pandas==1.4.2
Pillow==9.0.1
pip==21.2.4
opencv-python==4.6.0.66
PyYAML==6.0
matplotlib==3.5.1
seaborn==0.11.2
IPython==8.5.0
tqdm==4.64.1
